eliminate unnecessary module
